requirements.txt
pyyaml
aiohttp[speedups]
protobuf
lxml
beautifulsoup4
pycryptodome
aiomysql
opencv-contrib-python
